+++ src/common/lib/api/apiCollection/index.js
@@ -1,0 +1,8 @@
+export * from "./categories"
+export * from "./debio-balance"
+export * from "./emr"
+export * from "./labs"
+export * from "./location"
+export * from "./orders"
+export * from "./rating"
+export * from "./service-request"
