+++ src/views/Dashboard/Customer/Home/MyTest/index.vue
@@ -133,7 +133,7 @@ import localStorage from "@/common/lib/local-storage"
 import { u8aToHex } from "@polkadot/util"
 import { syncDecryptedFromIPFS } from "@/common/lib/ipfs"
 import { createSyncEvent } from "@/common/lib/ipfs/gcs"
-import { getCategories } from "@/common/lib/categories"
+import { getCategories } from "@/common/lib/api"
 import { getOrdersData } from "@/common/lib/polkadot-provider/query/orders"
 import { queryLabsById } from "@/common/lib/polkadot-provider/query/labs"
 import { queryServicesById } from "@/common/lib/polkadot-provider/query/services"
