+++ src/views/Dashboard/Customer/Home/PaymentHistory/Details.vue
@@ -82,8 +82,8 @@
 
 <script>
 import { alertIcon } from "@/common/icons"
-import { fetchPaymentDetails } from "@/common/lib/orders";
-import { getRatingService } from "@/common/lib/rating"
+import { fetchPaymentDetails } from "@/common/lib/api";
+import { getRatingService } from "@/common/lib/api"
 import { queryDnaSamples } from "@/common/lib/polkadot-provider/query/genetic-testing"
 import { mapState } from "vuex"
